brand-package/backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database & Storage
supabase==2.0.2
sqlalchemy==2.0.23
alembic==1.12.1

# AI & APIs
openai==1.3.7
httpx==0.24.1
aiohttp==3.9.0
pillow==10.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Rate Limiting & Caching
slowapi==0.1.9
python-redis==0.4.0

# Image Processing
cairosvg==2.7.1
svgwrite==1.4.3

# Background Tasks (optional)
celery==5.3.4
redis==5.0.1

# Utils
beautifulsoup4==4.12.2
dnspython==2.4.2
pandas==2.1.3
numpy==1.26.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Logging
loguru==0.7.2